64bit environment broken.
The fmlib crashed on 64b because the t_Device id was uint32_t and
the pointer stored within has 64b. The t_Device id parameters
was changed to uintptr_t type.

+++ include/fmd/ncsw_ext.h
@@ -408,7 +408,7 @@ typedef e_RxStoreResponse (t_RxStoreFunction)(t_Handle  h_App,
 
 
 typedef struct t_Device {
-    uint32_t    id;         /**< the device id */
+    uintptr_t   id;         /**< the device id */
     int         fd;         /**< the device file-descritpor */
     t_Handle    h_UserPriv;
     uint32_t    owners;
